requirements.txt
streamlit>=1.53.1
python-dotenv>=1.0.0
langchain>=1.2.0
langchain-core>=0.3.0
langchain-classic>=0.0.1
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.9.0
youtube-transcript-api>=0.6.0
wikipedia>=1.4.0
google-generativeai>=0.8.0
